+++ pyinjective/__init__.py
@@ -2,6 +2,6 @@
 # If this is not imported, importing later grpcio (required by AsyncClient) fails in Mac machines with M1 and M2
 from google.protobuf.internal import api_implementation  # noqa: F401
 
-from .async_client import AsyncClient  # noqa: F401
-from .transaction import Transaction  # noqa: F401
-from .wallet import Address, PrivateKey, PublicKey  # noqa: F401
+from pyinjective.async_client import AsyncClient  # noqa: F401
+from pyinjective.transaction import Transaction  # noqa: F401
+from pyinjective.wallet import Address, PrivateKey, PublicKey  # noqa: F401