+++ tests/client/chain/grpc/test_chain_grpc_exchange_api.py
@@ -30,6 +30,7 @@ class TestChainGrpcBankApi:
         spot_market_instant_listing_fee = coin_pb.Coin(denom="inj", amount="10000000000000000000")
         derivative_market_instant_listing_fee = coin_pb.Coin(denom="inj", amount="2000000000000000000000")
         binary_options_market_instant_listing_fee = coin_pb.Coin(denom="inj", amount="30000000000000000000")
+        admin = "inj1knhahceyp57j5x7xh69p7utegnnnfgxavmahjr"
         params = exchange_pb.Params(
             spot_market_instant_listing_fee=spot_market_instant_listing_fee,
             derivative_market_instant_listing_fee=derivative_market_instant_listing_fee,
@@ -56,6 +57,9 @@ class TestChainGrpcBankApi:
             minimal_protocol_fee_rate="0.000010000000000000",
             is_instant_derivative_market_launch_enabled=False,
             post_only_mode_height_threshold=57078000,
+            margin_decrease_price_timestamp_threshold_seconds=10,
+            exchange_admins=[admin],
+            inj_auction_max_cap="1000000000000000000000",
         )
         exchange_servicer.exchange_params.append(exchange_query_pb.QueryExchangeParamsResponse(params=params))
 
@@ -100,6 +104,11 @@ class TestChainGrpcBankApi:
                 "minimalProtocolFeeRate": params.minimal_protocol_fee_rate,
                 "isInstantDerivativeMarketLaunchEnabled": params.is_instant_derivative_market_launch_enabled,
                 "postOnlyModeHeightThreshold": str(params.post_only_mode_height_threshold),
+                "marginDecreasePriceTimestampThresholdSeconds": str(
+                    params.margin_decrease_price_timestamp_threshold_seconds
+                ),
+                "exchangeAdmins": [admin],
+                "injAuctionMaxCap": params.inj_auction_max_cap,
             }
         }
 
@@ -434,6 +443,9 @@ class TestChainGrpcBankApi:
             status=1,
             min_price_tick_size="0.000000000000001",
             min_quantity_tick_size="1000000000000000",
+            min_notional="5000000000000000000",
+            admin="inj1knhahceyp57j5x7xh69p7utegnnnfgxavmahjr",
+            admin_permissions=1,
         )
         exchange_servicer.spot_markets_responses.append(
             exchange_query_pb.QuerySpotMarketsResponse(
@@ -461,6 +473,9 @@ class TestChainGrpcBankApi:
                     "status": status_string,
                     "minPriceTickSize": market.min_price_tick_size,
                     "minQuantityTickSize": market.min_quantity_tick_size,
+                    "minNotional": market.min_notional,
+                    "admin": market.admin,
+                    "adminPermissions": market.admin_permissions,
                 }
             ]
         }
@@ -483,6 +498,9 @@ class TestChainGrpcBankApi:
             status=1,
             min_price_tick_size="0.000000000000001",
             min_quantity_tick_size="1000000000000000",
+            min_notional="5000000000000000000",
+            admin="inj1knhahceyp57j5x7xh69p7utegnnnfgxavmahjr",
+            admin_permissions=1,
         )
         exchange_servicer.spot_market_responses.append(
             exchange_query_pb.QuerySpotMarketResponse(
@@ -507,6 +525,9 @@ class TestChainGrpcBankApi:
                 "status": exchange_pb.MarketStatus.Name(market.status),
                 "minPriceTickSize": market.min_price_tick_size,
                 "minQuantityTickSize": market.min_quantity_tick_size,
+                "minNotional": market.min_notional,
+                "admin": market.admin,
+                "adminPermissions": market.admin_permissions,
             }
         }
 
@@ -528,6 +549,9 @@ class TestChainGrpcBankApi:
             status=1,
             min_price_tick_size="0.000000000000001",
             min_quantity_tick_size="1000000000000000",
+            min_notional="5000000000000000000",
+            admin="inj1knhahceyp57j5x7xh69p7utegnnnfgxavmahjr",
+            admin_permissions=1,
         )
         mid_price_and_tob = exchange_pb.MidPriceAndTOB(
             mid_price="2000000000000000000",
@@ -566,6 +590,9 @@ class TestChainGrpcBankApi:
                         "status": status_string,
                         "minPriceTickSize": market.min_price_tick_size,
                         "minQuantityTickSize": market.min_quantity_tick_size,
+                        "minNotional": market.min_notional,
+                        "admin": market.admin,
+                        "adminPermissions": market.admin_permissions,
                     },
                     "midPriceAndTob": {
                         "midPrice": mid_price_and_tob.mid_price,
@@ -594,6 +621,9 @@ class TestChainGrpcBankApi:
             status=1,
             min_price_tick_size="0.000000000000001",
             min_quantity_tick_size="1000000000000000",
+            min_notional="5000000000000000000",
+            admin="inj1knhahceyp57j5x7xh69p7utegnnnfgxavmahjr",
+            admin_permissions=1,
         )
         mid_price_and_tob = exchange_pb.MidPriceAndTOB(
             mid_price="2000000000000000000",
@@ -630,6 +660,9 @@ class TestChainGrpcBankApi:
                     "status": status_string,
                     "minPriceTickSize": market.min_price_tick_size,
                     "minQuantityTickSize": market.min_quantity_tick_size,
+                    "minNotional": market.min_notional,
+                    "admin": market.admin,
+                    "adminPermissions": market.admin_permissions,
                 },
                 "midPriceAndTob": {
                     "midPrice": mid_price_and_tob.mid_price,
@@ -698,6 +731,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.trader_spot_orders_responses.append(
             exchange_query_pb.QueryTraderSpotOrdersResponse(
@@ -719,6 +753,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -736,6 +771,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.account_address_spot_orders_responses.append(
             exchange_query_pb.QueryAccountAddressSpotOrdersResponse(
@@ -757,6 +793,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -774,6 +811,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.spot_orders_by_hashes_responses.append(
             exchange_query_pb.QuerySpotOrdersByHashesResponse(
@@ -796,6 +834,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -811,6 +850,7 @@ class TestChainGrpcBankApi:
             price="1000000000000000000",
             quantity="1000000000000000",
             isReduceOnly=False,
+            cid="buy_cid",
         )
         buy_order = exchange_pb.SubaccountOrderData(
             order=buy_subaccount_order,
@@ -820,6 +860,7 @@ class TestChainGrpcBankApi:
             price="2000000000000000000",
             quantity="2000000000000000",
             isReduceOnly=False,
+            cid="sell_cid",
         )
         sell_order = exchange_pb.SubaccountOrderData(
             order=sell_subaccount_order,
@@ -845,6 +886,7 @@ class TestChainGrpcBankApi:
                         "price": buy_subaccount_order.price,
                         "quantity": buy_subaccount_order.quantity,
                         "isReduceOnly": buy_subaccount_order.isReduceOnly,
+                        "cid": buy_subaccount_order.cid,
                     },
                     "orderHash": base64.b64encode(buy_order.order_hash).decode(),
                 }
@@ -855,6 +897,7 @@ class TestChainGrpcBankApi:
                         "price": sell_subaccount_order.price,
                         "quantity": sell_subaccount_order.quantity,
                         "isReduceOnly": sell_subaccount_order.isReduceOnly,
+                        "cid": sell_subaccount_order.cid,
                     },
                     "orderHash": base64.b64encode(sell_order.order_hash).decode(),
                 }
@@ -874,6 +917,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.trader_spot_transient_orders_responses.append(
             exchange_query_pb.QueryTraderSpotOrdersResponse(
@@ -895,6 +939,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -1007,6 +1052,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.trader_derivative_orders_responses.append(
             exchange_query_pb.QueryTraderDerivativeOrdersResponse(
@@ -1029,6 +1075,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -1047,6 +1094,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.account_address_derivative_orders_responses.append(
             exchange_query_pb.QueryAccountAddressDerivativeOrdersResponse(
@@ -1069,6 +1117,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -1087,6 +1136,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.derivative_orders_by_hashes_responses.append(
             exchange_query_pb.QueryDerivativeOrdersByHashesResponse(
@@ -1110,6 +1160,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -1128,6 +1179,7 @@ class TestChainGrpcBankApi:
             fillable="1000000000000000",
             isBuy=True,
             order_hash="0x14e43adbb3302db28bcd0619068227ebca880cdd66cdfc8b4a662bcac0777849",
+            cid="order_cid",
         )
         exchange_servicer.trader_derivative_transient_orders_responses.append(
             exchange_query_pb.QueryTraderDerivativeOrdersResponse(
@@ -1150,6 +1202,7 @@ class TestChainGrpcBankApi:
                     "fillable": order.fillable,
                     "isBuy": order.isBuy,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }
@@ -1178,6 +1231,9 @@ class TestChainGrpcBankApi:
             status=1,
             min_price_tick_size="100000000000000000000",
             min_quantity_tick_size="1000000000000000",
+            min_notional="5000000000000000000",
+            admin="inj1knhahceyp57j5x7xh69p7utegnnnfgxavmahjr",
+            admin_permissions=1,
         )
         market_info = exchange_pb.PerpetualMarketInfo(
             market_id="0x17ef48032cb24375ba7c2e39f384e56433bcab20cbee9a7357e4cba2eb00abe6",
@@ -1239,6 +1295,9 @@ class TestChainGrpcBankApi:
                         "status": status_string,
                         "minPriceTickSize": market.min_price_tick_size,
                         "minQuantityTickSize": market.min_quantity_tick_size,
+                        "minNotional": market.min_notional,
+                        "admin": market.admin,
+                        "adminPermissions": market.admin_permissions,
                     },
                     "perpetualInfo": {
                         "marketInfo": {
@@ -1288,6 +1347,9 @@ class TestChainGrpcBankApi:
             status=1,
             min_price_tick_size="100000000000000000000",
             min_quantity_tick_size="1000000000000000",
+            min_notional="5000000000000000000",
+            admin="inj1knhahceyp57j5x7xh69p7utegnnnfgxavmahjr",
+            admin_permissions=1,
         )
         market_info = exchange_pb.PerpetualMarketInfo(
             market_id="0x17ef48032cb24375ba7c2e39f384e56433bcab20cbee9a7357e4cba2eb00abe6",
@@ -1347,6 +1409,9 @@ class TestChainGrpcBankApi:
                     "status": status_string,
                     "minPriceTickSize": market.min_price_tick_size,
                     "minQuantityTickSize": market.min_quantity_tick_size,
+                    "minNotional": market.min_notional,
+                    "admin": market.admin,
+                    "adminPermissions": market.admin_permissions,
                 },
                 "perpetualInfo": {
                     "marketInfo": {
@@ -2256,6 +2321,8 @@ class TestChainGrpcBankApi:
             min_price_tick_size="100000000000000000000",
             min_quantity_tick_size="1000000000000000",
             settlement_price="2000000000000000000",
+            min_notional="5000000000000000000",
+            admin_permissions=1,
         )
         response = exchange_query_pb.QueryBinaryMarketsResponse(
             markets=[market],
@@ -2285,6 +2352,8 @@ class TestChainGrpcBankApi:
                     "minPriceTickSize": market.min_price_tick_size,
                     "minQuantityTickSize": market.min_quantity_tick_size,
                     "settlementPrice": market.settlement_price,
+                    "minNotional": market.min_notional,
+                    "adminPermissions": market.admin_permissions,
                 },
             ]
         }
@@ -2304,6 +2373,7 @@ class TestChainGrpcBankApi:
             isBuy=True,
             isLimit=True,
             order_hash="0x17ef48032cb24375ba7c2e39f384e56433bcab20cbee9a7357e4cba2eb00abe6",
+            cid="order_cid",
         )
         response = exchange_query_pb.QueryTraderDerivativeConditionalOrdersResponse(orders=[order])
         exchange_servicer.trader_derivative_conditional_orders_responses.append(response)
@@ -2324,6 +2394,7 @@ class TestChainGrpcBankApi:
                     "isBuy": order.isBuy,
                     "isLimit": order.isLimit,
                     "orderHash": order.order_hash,
+                    "cid": order.cid,
                 }
             ]
         }