+++ tests/client/indexer/stream_grpc/test_indexer_grpc_spot_stream.py
@@ -54,6 +54,7 @@ class TestIndexerGrpcSpotStream:
             service_provider_fee="0.4",
             min_price_tick_size="0.000000000000001",
             min_quantity_tick_size="1000000000000000",
+            min_notional="1000000",
         )
 
         spot_servicer.stream_markets_responses.append(
@@ -109,6 +110,7 @@ class TestIndexerGrpcSpotStream:
                 "serviceProviderFee": market.service_provider_fee,
                 "minPriceTickSize": market.min_price_tick_size,
                 "minQuantityTickSize": market.min_quantity_tick_size,
+                "minNotional": market.min_notional,
             },
             "operationType": operation_type,
             "timestamp": str(timestamp),