+++ tests/rpc_fixtures/markets_fixtures.py
@@ -29,9 +29,9 @@ def smart_denom_metadata():
 
 @pytest.fixture
 def inj_token_meta():
-    from pyinjective.proto.exchange.injective_spot_exchange_rpc_pb2 import TokenMeta
+    from pyinjective.proto.exchange import injective_spot_exchange_rpc_pb2 as spot_exchange_pb
 
-    token = TokenMeta(
+    token = spot_exchange_pb.TokenMeta(
         name="Injective Protocol",
         address="0xe28b3B32B6c345A34Ff64674606124Dd5Aceca30",
         symbol="INJ",
@@ -45,9 +45,9 @@ def inj_token_meta():
 
 @pytest.fixture
 def ape_token_meta():
-    from pyinjective.proto.exchange.injective_spot_exchange_rpc_pb2 import TokenMeta
+    from pyinjective.proto.exchange import injective_spot_exchange_rpc_pb2 as spot_exchange_pb
 
-    token = TokenMeta(
+    token = spot_exchange_pb.TokenMeta(
         name="APE",
         address="0x0000000000000000000000000000000000000000",
         symbol="APE",
@@ -61,9 +61,9 @@ def ape_token_meta():
 
 @pytest.fixture
 def usdt_token_meta():
-    from pyinjective.proto.exchange.injective_spot_exchange_rpc_pb2 import TokenMeta
+    from pyinjective.proto.exchange import injective_spot_exchange_rpc_pb2 as spot_exchange_pb
 
-    token = TokenMeta(
+    token = spot_exchange_pb.TokenMeta(
         name="USDT",
         address="0x0000000000000000000000000000000000000000",
         symbol="USDT",
@@ -77,9 +77,9 @@ def usdt_token_meta():
 
 @pytest.fixture
 def usdt_token_meta_second_denom():
-    from pyinjective.proto.exchange.injective_spot_exchange_rpc_pb2 import TokenMeta
+    from pyinjective.proto.exchange import injective_spot_exchange_rpc_pb2 as spot_exchange_pb
 
-    token = TokenMeta(
+    token = spot_exchange_pb.TokenMeta(
         name="USDT Second Denom",
         address="0x0000000000000000000000000000000000000000",
         symbol="USDT",
@@ -93,9 +93,9 @@ def usdt_token_meta_second_denom():
 
 @pytest.fixture
 def usdt_perp_token_meta():
-    from pyinjective.proto.exchange.injective_derivative_exchange_rpc_pb2 import TokenMeta
+    from pyinjective.proto.exchange import injective_derivative_exchange_rpc_pb2 as derivative_exchange_pb
 
-    token = TokenMeta(
+    token = derivative_exchange_pb.TokenMeta(
         name="Tether",
         address="0xdAC17F958D2ee523a2206206994597C13D831ec7",
         symbol="USDTPerp",
@@ -109,9 +109,9 @@ def usdt_perp_token_meta():
 
 @pytest.fixture
 def ape_usdt_spot_market_meta(ape_token_meta, usdt_token_meta_second_denom):
-    from pyinjective.proto.exchange.injective_spot_exchange_rpc_pb2 import SpotMarketInfo
+    from pyinjective.proto.exchange import injective_spot_exchange_rpc_pb2 as spot_exchange_pb
 
-    market = SpotMarketInfo(
+    market = spot_exchange_pb.SpotMarketInfo(
         market_id="0x8b67e705bb4e09c88aecfc295569481dbf2fe1d5efe364651fbe72385938e000",
         market_status="active",
         ticker="APE/USDT",
@@ -131,9 +131,9 @@ def ape_usdt_spot_market_meta(ape_token_meta, usdt_token_meta_second_denom):
 
 @pytest.fixture
 def inj_usdt_spot_market_meta(inj_token_meta, usdt_token_meta):
-    from pyinjective.proto.exchange.injective_spot_exchange_rpc_pb2 import SpotMarketInfo
+    from pyinjective.proto.exchange import injective_spot_exchange_rpc_pb2 as spot_exchange_pb
 
-    market = SpotMarketInfo(
+    market = spot_exchange_pb.SpotMarketInfo(
         market_id="0x7a57e705bb4e09c88aecfc295569481dbf2fe1d5efe364651fbe72385938e9b0",
         market_status="active",
         ticker="INJ/USDT",
@@ -153,25 +153,21 @@ def inj_usdt_spot_market_meta(inj_token_meta, usdt_token_meta):
 
 @pytest.fixture
 def btc_usdt_perp_market_meta(usdt_perp_token_meta):
-    from pyinjective.proto.exchange.injective_derivative_exchange_rpc_pb2 import (
-        DerivativeMarketInfo,
-        PerpetualMarketFunding,
-        PerpetualMarketInfo,
-    )
+    from pyinjective.proto.exchange import injective_derivative_exchange_rpc_pb2 as derivative_exchange_pb
 
-    perpetual_market_info = PerpetualMarketInfo(
+    perpetual_market_info = derivative_exchange_pb.PerpetualMarketInfo(
         hourly_funding_rate_cap="0.0000625",
         hourly_interest_rate="0.00000416666",
         next_funding_timestamp=1684764000,
         funding_interval=3600,
     )
-    perpetual_market_funding = PerpetualMarketFunding(
+    perpetual_market_funding = derivative_exchange_pb.PerpetualMarketFunding(
         cumulative_funding="6880500093.266083891331674194",
         cumulative_price="-0.952642601240470199",
         last_timestamp=1684763442,
     )
 
-    market = DerivativeMarketInfo(
+    market = derivative_exchange_pb.DerivativeMarketInfo(
         market_id="0x4ca0f92fc28be0c9761326016b5a1a2177dd6375558365116b5bdda9abc229ce",
         market_status="active",
         ticker="BTC/USDT PERP",
@@ -198,9 +194,9 @@ def btc_usdt_perp_market_meta(usdt_perp_token_meta):
 
 @pytest.fixture
 def first_match_bet_market_meta(inj_usdt_spot_market_meta):
-    from pyinjective.proto.exchange.injective_derivative_exchange_rpc_pb2 import BinaryOptionsMarketInfo
+    from pyinjective.proto.exchange import injective_derivative_exchange_rpc_pb2 as derivative_exchange_pb
 
-    market = BinaryOptionsMarketInfo(
+    market = derivative_exchange_pb.BinaryOptionsMarketInfo(
         market_id="0x230dcce315364ff6360097838701b14713e2f4007d704df20ed3d81d09eec957",
         market_status="active",
         ticker="5fdbe0b1-1707800399-WAS",